requirements.txt
faiss_cpu==1.8.0.post1
Flask==3.0.3
Flask_Cors==5.0.0
mysql-connector-python==9.0.0
numpy==1.24.4
sentence-transformers==3.2.1
spark-ai-python==0.4.5
